code split between app.js and index.js
